Ensure that the members of sg_source are initialized properly in sg_source_anon_init().
There was a chance that during sg_source_acquire() a failure of one of the SG*() functions would lead to sg_source_release() being called with non-null members of the sg_source structure. Specifically, the gworld pointer would be non-null and then subsequently UnlockPixels would be called on a garbage pointer.