Merge Chris Cooksey's branches/framesize_enhancements back into trunk.
Added support for DV cameras. In addition, added support for any frame size,
any frame rate, or any of three color spaces requested by the client,
regardless of whether they are physically supported by the device.

NOTE: these enhancements have NOT been added to the Linux source.

Also note that device aspect ratios are not preserved. If a 4:3 frame size
is requested from a device that has a 16:9 resolution, the image will be
scaled to fit.

These changes do NOT alter the public interface.